+++ validation/kubernetes/prometheus/main.tf
@@ -12,9 +12,8 @@ module "prometheus" {
     helm       = helm
   }
 
-  remote_write_url              = "https://my-remote-writer.com"
-  cloud_provider                = "azure"
-  cluster_name                  = "aks1"
-  environment                   = "dev"
-  kube_state_metrics_namespaces = ""
+  remote_write_url = "https://my-remote-writer.com"
+  cloud_provider   = "azure"
+  cluster_name     = "aks1"
+  environment      = "dev"
 }